+++ src/migration/generator.rs
@@ -160,7 +160,7 @@ pub fn generate_initial_migration(
         tables: tables.values().cloned().collect(),
         edges: edges.values().cloned().collect(),
         buckets: buckets.values().cloned().collect(),
-        analyzers: Vec::new(),
+        ..Default::default()
     };
 
     let (up_statements, down_statements) = build_initial_statements(&snapshot)?;
@@ -249,6 +249,7 @@ pub fn create_blank_migration(
 ///     event: None,
 ///     bucket: None,
 ///     analyzer: None,
+///     object: None,
 ///     description: "Add user table".into(),
 ///     forward_sql: "DEFINE TABLE user SCHEMAFULL;".into(),
 ///     backward_sql: "REMOVE TABLE user;".into(),
@@ -940,6 +941,7 @@ mod tests {
             event: None,
             bucket: None,
             analyzer: None,
+            object: None,
             description: format!("Add {name} table"),
             forward_sql: format!("DEFINE TABLE {name} SCHEMAFULL;"),
             backward_sql: format!("REMOVE TABLE {name};"),
@@ -987,6 +989,7 @@ mod tests {
                 event: None,
                 bucket: None,
                 analyzer: None,
+                object: None,
                 description: "x".into(),
                 forward_sql: String::new(),
                 backward_sql: String::new(),
@@ -1013,6 +1016,7 @@ mod tests {
             event: None,
             bucket: None,
             analyzer: None,
+            object: None,
             description: "x".into(),
             forward_sql: "DEFINE TABLE x SCHEMAFULL".into(),
             backward_sql: "REMOVE TABLE x".into(),